Extract MessageMappingForm to its own component
